branch start, combine both fuzz targets into one
