+++ .github/workflows/ci.yml
@@ -72,8 +72,8 @@ jobs:
         run: |
           cargo miri test --all-features --manifest-path=compact_str/Cargo.toml
 
-  fuzz-creation:
-    name: fuzz creation
+  fuzz:
+    name: fuzz
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
@@ -91,28 +91,7 @@ jobs:
           version: latest
       - name: Fuzz!
         run: |
-          cargo fuzz run --debug-assertions creation -- -max_total_time=120
-
-  fuzz-modify:
-    name: fuzz modify
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-        name: Checkout repo
-      - uses: actions-rs/toolchain@v1
-        name: Install Rust and friends
-        with:
-          profile: minimal
-          toolchain: nightly
-          override: true
-      - uses: actions-rs/install@v0.1
-        name: Install cargo-fuzz
-        with:
-          crate: cargo-fuzz
-          version: latest
-      - name: Fuzz!
-        run: |
-          cargo fuzz run --debug-assertions modify -- -max_total_time=120
+          cargo fuzz run --debug-assertions compact_str -- -max_total_time=120
 
   example-bytes:
     name: example - bytes
